Implement zkapp transactions
Squashed commit of the following:

commit bef440b0a8de0de6f9859c5e720e3bb9be0a08bb
Author: Daniel Kuehr <daniel.kuehr@viablesystems.io>
Date:   Mon Jan 16 18:52:05 2023 -0300

    remove duplicated code, and merge transaction implementation into scan_state

commit 8de21583632752eaa6683b18af10e7d83f2a83af
Author: Sebastien Chapuis <sebastien@chapu.is>
Date:   Tue Jan 10 17:42:43 2023 +0100

    Make `transaction::apply_transaction` generic over the ledger

    Transactions are applied on `LedgerIntf`

commit c8fb141b22e2967346398d0801f6840eeb58dcc5
Author: Sebastien Chapuis <sebastien@chapu.is>
Date:   Tue Jan 10 13:27:59 2023 +0100

    Use `ExistingOrNew` instead of `GetOrCreated`

    OCaml uses `ExistingOrNew`

commit 81d28e3afce4ad31dd5322dfb6677c3408acb058
Author: Sebastien Chapuis <sebastien@chapu.is>
Date:   Tue Jan 10 13:22:53 2023 +0100

    Use `hash::Inputs` instead of `Chunked`

commit 064564a6e6e04467c62f76bdd5802b4cac1ddd2c
Author: Sebastien Chapuis <sebastien@chapu.is>
Date:   Tue Jan 10 10:31:17 2023 +0100

    Merge branch 'transaction_fuzzer' into merge_tx_fuzzer

commit f73e37b7a7fbc46d4e68f3cc707bbac5e5f7b8b1
Author: Daniel Kuehr <daniel.kuehr@viablesystems.io>
Date:   Sat Jan 7 01:18:02 2023 -0300

    WIP: rebase against main, resolve conflicts and merge duplicated code

commit 86d0d79a8cc7e8be92d29e24c58202cfbe8529be
Author: Daniel Kuehr <daniel.kuehr@viablesystems.io>
Date:   Tue Jan 3 20:38:39 2023 -0300

    WIP: zkapp transaction port

commit 6435a10759864ba49ce3db696124d44ef841015a
Author: Daniel Kuehr <daniel.kuehr@viablesystems.io>
Date:   Sun Dec 18 21:17:34 2022 -0300

    implement user command transactions

commit 786028f9a0653778a8febc843e774c56c915b356
Author: Daniel Kuehr <daniel.kuehr@viablesystems.io>
Date:   Thu Dec 15 22:59:10 2022 -0300

    WIP: port OCaml transaction logic into Rust

commit f3003102564744dfa37650ffbcf77464ccf7a504
Author: Daniel Kuehr <daniel.kuehr@viablesystems.io>
Date:   Thu Dec 15 22:58:18 2022 -0300

    fix total_currency type to u64

+++ ledger/src/blocks/protocol_state.rs
@@ -47,7 +47,7 @@ pub struct StagedLedgerHash {
     pub pending_coinbase_hash: Fp,
 }
 
-#[derive(Clone, Debug)]
+#[derive(Clone, PartialEq, Debug)]
 pub enum Sgn {
     Pos,
     Neg,
@@ -91,7 +91,7 @@ pub struct ConsensusGlobalSlot {
 #[derive(Clone, Debug)]
 pub struct EpochLedger {
     pub hash: Fp,
-    pub total_currency: i64,
+    pub total_currency: u64,
 }
 
 #[derive(Clone, Debug)]
@@ -110,7 +110,7 @@ pub struct ConsensusState {
     pub min_window_density: u32,
     pub sub_window_densities: Vec<u32>,
     pub last_vrf_output: [u8; 32], // TODO: In binprot it's a string ?
-    pub total_currency: i64,
+    pub total_currency: u64,
     pub curr_global_slot: ConsensusGlobalSlot,
     pub global_slot_since_genesis: u32,
     pub staking_epoch_data: DataStaking,
@@ -215,7 +215,7 @@ impl ProtocolStateBody {
                     inputs.append_bool(last_byte & bit != 0);
                 }
             }
-            inputs.append_u64(consensus.total_currency as u64);
+            inputs.append_u64(consensus.total_currency);
             inputs.append_u32(consensus.curr_global_slot.slot_number);
             inputs.append_u32(consensus.curr_global_slot.slots_per_epoch);
             inputs.append_u32(consensus.global_slot_since_genesis);
@@ -226,7 +226,7 @@ impl ProtocolStateBody {
                 inputs.append_field(data.start_checkpoint);
                 inputs.append_u32(data.epoch_length);
                 inputs.append_field(data.ledger.hash);
-                inputs.append_u64(data.ledger.total_currency as u64);
+                inputs.append_u64(data.ledger.total_currency);
                 inputs.append_field(data.lock_checkpoint);
             }
             inputs.append_field(consensus.block_stake_winner.x);
